First steps towards 64-bit build
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk@1307 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ pwsafe/pwsafe/corelib/PWScore.cpp
@@ -1150,7 +1150,7 @@ PWScore::ImportKeePassTextFile(const CMyString &filename)
     // the first line of the keepass text file contains a few garbage characters
     linebuf = linebuf.erase(0, linebuf.find("["));
 
-    int pos = -1;
+    size_t pos = static_cast<size_t>(-1);
     for (;;) {
         if (!ifs)
             break;
